+++ src/Components/CollabZone/CollabFeatures/JoinUs.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import MainSection from "./MainSection";
 import FormModal from "./FormModal";
-import FormSteps from "./FormSteps";
+import AdvancedForm from "./JoinCollabForm/AdvancedForm";
 
 const JoinUs = () => {
   const [isFormVisible, setIsFormVisible] = useState(false);
@@ -14,7 +14,7 @@ const JoinUs = () => {
     <>
       <MainSection toggleFormVisibility={toggleFormVisibility} />
       <FormModal isVisible={isFormVisible} closeForm={closeForm}>
-        <FormSteps closeForm={closeForm} />
+              <AdvancedForm closeForm={closeForm} />
       </FormModal>
     </>
   );
